Revert "Divine"
