+++ models/userModel.js
@@ -25,7 +25,7 @@ const validation = (data)=>{
         firstName: joi.string().required().min(4),
         lastName: joi.string().required().min(4),
         email: joi.string().required().email(),
-        password: joi.string().required().email()
+        password: joi.string().required().email().min(8)
     })
     return data.validate()
 }
