+++ package.json
@@ -27,7 +27,6 @@
     "lodash": "^4.17.21",
     "mongoose": "^6.5.2",
     "nodemailer": "^6.7.8",
-    "nodemailer-smtp-transport": "^2.7.4",
     "nodemon": "^2.0.19",
     "otp-generator": "^4.0.0"
   }
