--- routes/user.js
@@ -1,8 +1,0 @@
-const { signup } = require('../contollers/userInfo.js')
-const { verifyToken } = require('../auth/user')
-const express = require('express')
-
-const router = express.Router()
-router.post('/signup', signup())  
-
-module.exports = router;
