+++ server.js
@@ -1,11 +1,9 @@
 const dotenv = require('dotenv')
 dotenv.config()
-// const {Router} = require('./routes/user.js')
 const express = require('express')
 const app = express()
 
 const mongoose = require('mongoose')
-const router = require('./routes/user.js')
 
 mongoose.connect(process.env.URL).then(()=>{
     console.log("Database successfully connected");
@@ -15,7 +13,7 @@ mongoose.connect(process.env.URL).then(()=>{
 const PORT = process.env.PORT
 app.use(express.json())
 app.use(express.urlencoded({extended: true}))
-app.use("/api/auth",router)
+
 
 app.listen(PORT, ()=>{
     console.log(`The server is learning on port ${PORT}`)
